% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agraph.R
\name{admixture_edge}
\alias{admixture_edge}
\title{Create admixture edges from child to two parents}
\usage{
admixture_edge(child, parent1, parent2)
}
\arguments{
\item{child}{The name of the child node}

\item{parent1}{The name of the parent node}

\item{parent2}{The name of the parent node}
}
\description{
Syntactic suggar for constructing edges in an admixture graph.
}

