% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agraph.R
\name{agraph}
\alias{agraph}
\title{Create an admixture graph object.}
\usage{
agraph(leaves, inner_nodes, parent_edges, admixture_proportions)
}
\arguments{
\item{leaves}{The names of the leaves in the admixture graph.}

\item{inner_nodes}{The name of the inner nodes in the admxture graph.}

\item{parent_edges}{The list of edges in the graph, created by
  \code{\link{parent_edges}}.}

\item{admixture_proportions}{The list of admixture proportions; created by
  \code{\link{admixture_proportions}}}
}
\value{
An admixture graph object.
}
\description{
Create an admixture graph object.
}
\examples{
leaves <- c("A", "B", "C", "O")
inner_nodes <- c("AB", "BC", "ABC", "R")
edges <- parent_edges(c(edge("A", "AB"),
                        edge("C", "BC"),
                        admixture_edge("B", "AB", "BC"),
                        edge("AB", "ABC"),
                        edge("BC", "ABC"),
                        edge("ABC", "R"),
                        edge("O", "R")))
admixtures <- admixture_proportions(c(admix_props("B", "AB", "BC", "a")))

graph <- agraph(leaves, inner_nodes, edges, admixtures)
}
\seealso{
edge

admixture_edge

parent_edges

admix_props

admixture_proportions
}

