% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agraph.R
\name{agraph_parents}
\alias{agraph_parents}
\title{Build a parent incidence matrix from an edge list.}
\usage{
agraph_parents(nodes, parent_edges)
}
\arguments{
\item{nodes}{The name of the nodes in the admxture graph.}

\item{parent_edges}{An \eqn{n \times 2} matrix where the first column is the
child node and the second the parent.}
}
\value{
An incidence matrix for the parent structure of an admixture graph.
}
\description{
Build a parent incidence matrix from an edge list.
}

