% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose.R
\name{all_graphs}
\alias{all_graphs}
\title{All graphs.}
\usage{
all_graphs(populations, admixture_events)
}
\arguments{
\item{populations}{A vector of populations (leaf names).}

\item{admixture_events}{The maximum number of admixture events allowed.}
}
\value{
A list of admixture graphs.
}
\description{
Gives a list of all the graphs with at most a given number of admixture events.
No duplicates.
}
\seealso{
\code{\link{fit_graph_list}}
}

