% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{all_path_overlaps}
\alias{all_path_overlaps}
\title{Get the list of overlaps of all paths.}
\usage{
all_path_overlaps(paths1, paths2)
}
\arguments{
\item{paths1}{Paths between one pair of leaves.}

\item{paths2}{Paths between another pair of leaves.}
}
\value{
A list of the overlaps of all combinations of paths from \code{paths1}
        and \code{paths2}.
}
\description{
Gets the list of overlaps of all paths.
}

