% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{all_paths}
\alias{all_paths}
\title{Compute all paths from one leaf to another.}
\usage{
all_paths(graph, src, dst)
}
\arguments{
\item{graph}{The admixture graph.}

\item{src}{The starting leaf.}

\item{dst}{The destination leaf.}
}
\value{
A list containing all the paths from \code{src} to \code{dst}.
}
\description{
Computes all paths from one leaf to another.
}

