% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bears}
\alias{bears}
\title{Statistics for populations of bears}
\format{A data frame with 18 rows and 6 variables:
\describe{
 \item{W}{The W population}
 \item{X}{The X population}
 \item{Y}{The Y population}
 \item{Z}{The Z population}
 \item{D}{The D (f_4(W,X;Y,Z)) statistics}
 \item{Z.value}{The blocked jacknife Z values}
}}
\source{
\url{http://onlinelibrary.wiley.com/doi/10.1111/mec.13038/abstract}
}
\usage{
bears
}
\description{
Computed f_4(W,X;Y,Z) statistics for different populations of bears.
}
\keyword{datasets}

