% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{build_edge_optimisation_matrix}
\alias{build_edge_optimisation_matrix}
\title{Build a matrix coding the linear system of edges once the admix variables
have been fixed.}
\usage{
build_edge_optimisation_matrix(data, graph,
  parameters = extract_graph_parameters(graph))
}
\arguments{
\item{data}{The data set.}

\item{graph}{The admixture graph.}

\item{parameters}{In case one wants to tweak something in the graph.}
}
\value{
A list containing the full matrix ($full),a version with zero columns
        removed ($column_reduced) and parameters to pass forward ($parameters).
}
\description{
The elements are characters containing numerals, admix variable names,
parenthesis and arithmetical operations. (Transform into expressions with
parse and then evaluate with eval). The default column names are the edge names
from extract_graph_parameter$edges, the rows have no names.
}

