% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{burn_in}
\alias{burn_in}
\title{Removes the first k rows from a trace.}
\usage{
burn_in(trace, k)
}
\arguments{
\item{trace}{A trace from an MCMC run.}

\item{k}{Number of rows to discard as burn-in.}
}
\description{
Removes the first k rows from a trace.
}

