% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{canonise_expression}
\alias{canonise_expression}
\title{Used to recognize similar expressions and to possibly simplify them.}
\usage{
canonise_expression(x)
}
\arguments{
\item{x}{Input is assumed to be a char containing numerals, variables and
           clauses \code{(1 - variable)} (mind the spaces, this is how the
           f4-function outputs), separated by \code{+}, \code{-} or \code{*},
           with no spaces.
           Everything is pretty much ruined if variable names contain forbidden
           symbols \code{+, -, *, (, )} or are purely numerals.}
}
\value{
A polynomial in a canonical form with no parenthesis or spaces and the
          monomials in lexicographical order.
}
\description{
This is not pretty but let's see if it speeds up the program.
}

