% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose.R
\name{canonise_graph}
\alias{canonise_graph}
\title{Canonise graph.}
\usage{
canonise_graph(graph)
}
\arguments{
\item{graph}{An admixture graph.}
}
\value{
A logical vector coding the parental matrix of a canonised version of the input graph.
}
\description{
Given a graph builds a unique logical vector depending only on the leaf names and the graph
topology (not the inner node names, root position or the input order of edges or inner nodes).
Can be used to detect graph isomorphism, that is, to weed out duplicates from a graph list.
Only for comparison of graphs with the same leaf set!
}

