% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{coef.agraph_fit}
\alias{coef.agraph_fit}
\title{Get fitted parameters for a fitted graph.}
\usage{
\method{coef}{agraph_fit}(object, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{...}{Additional parameters.}
}
\description{
Extract the graph parameters for a graph fitted to data.
}

