% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agraph.R
\name{edge}
\alias{edge}
\title{Create an edge from a child to a parent.}
\usage{
edge(child, parent)
}
\arguments{
\item{child}{The name of the child node.}

\item{parent}{The name of the parent node.}
}
\description{
Syntactic suggar for constructing edges in an admixture graph.
}

