% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{edge_optimisation_function}
\alias{edge_optimisation_function}
\title{More detailed edge fitting than mere \code{cost_function}.}
\usage{
edge_optimisation_function(data, matrix, graph,
  parameters = extract_graph_parameters(graph))
}
\arguments{
\item{data}{The data set.}

\item{matrix}{A full  edge optimisation matrix (typically given by the
function \code{edge_optimisation_matrix$full}).}

\item{graph}{The admixture graph.}

\item{parameters}{In case one wants to tweak something in the graph.}
}
\value{
Given an input vector of admix variables, returns a list \code{x} containing
         the minimal error (\code{x$cost}), the graph-f4-statistics
         (\code{x$approximation}), an example solution (\code{x$edge_fit}), linear
         relations describing all the solutions (\code{x$homogeneous}) and one
         way to choose the free (\code{x$free_edges}) and bounded
         (\code{x$bounded_edges}) edge variables.
}
\description{
Returning the cost, an example edge solution of an optimal fit, and linear
relations describing the set of all edge solutions. Operating with the full
edge optimisation matrix.
}

