% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-f-statistics.R
\name{evaluate_f4}
\alias{evaluate_f4}
\title{Evaluates an f_4 statistics in a given environment.}
\usage{
evaluate_f4(graph, env, W, X, Y, Z)
}
\arguments{
\item{graph}{The admixture graph.}

\item{env}{The environment containing the graph parameters.}

\item{W}{First population/sample.}

\item{X}{Second population/sample.}

\item{Y}{Third population/sample.}

\item{Z}{Fourth population/sample.}
}
\value{
The \eqn{f_4} value specified by the graph and the environment.
}
\description{
Evaluates an \eqn{f_4} statistics in a given environment.
}

