% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_to_qpGraph}
\alias{export_to_qpGraph}
\title{Export to Patterson's qpGraph format.}
\usage{
export_to_qpGraph(f, graph)
}
\arguments{
\item{f}{File object, e.g. stdout()}

\item{graph}{A graph to export.}
}
\description{
This function writes a graph to a file-object, f, in the format used by the
qpGraph tool. This format takes admixture proportions as part of the specification,
but since we do not hold these proportions in our graphs, but only in fitted data,
the export functions puts all admixture proportions as 50%/50%. Edit the output
file by hand if you want to change this.
}

