% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agraph.R
\name{extract_admixture_proportion_parameters}
\alias{extract_admixture_proportion_parameters}
\title{Extract the admixture proportion parameter from edge specifications.}
\usage{
extract_admixture_proportion_parameters(parent_edges)
}
\arguments{
\item{parent_edges}{Matrix created with the \code{agraph_parents} function.}
}
\value{
The parents edges reduced to the rows with admixture proportions.
}
\description{
This function is simply selecting the edges with admixture proportion specifications
so these can be handled when building a graph using \code{agraph}. It is not a function
you would need to call explicitly, rather it is there to allow people \emph{not} to use
it to provide admixture proportions explicitly (which we normally wouldn't recommend).
}

