% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agraph.R
\name{extract_trees}
\alias{extract_trees}
\title{Extract trees}
\usage{
extract_trees(graph)
}
\arguments{
\item{graph}{An agraph object}
}
\value{
A list of trees
}
\description{
Extracts all the trees embedded in an agraph object
}

