% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f-statistics.R
\name{f2}
\alias{f2}
\title{Calculate the f_2(A, B) statistics.}
\usage{
f2(graph, A, B)
}
\arguments{
\item{graph}{The admixture graph.}

\item{A}{A leaf node.}

\item{B}{A leaf node.}
}
\value{
A symbolic representation of the equation for the \eqn{f_2}
        statistics given by the admixture graph.
}
\description{
Calculate the \eqn{f_2(A, B)} statistics.
}

