% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f4_plotting.R
\name{f4stats}
\alias{f4stats}
\title{Make a data frame an f4 statistics object.}
\usage{
f4stats(x)
}
\arguments{
\item{x}{Data frame with observed D (f4) statistics}
}
\description{
This is mostly just a convinience function to set the class of a data frame such that
we plot data as error bars in a meaningful way for statistics for admixture graphs.
}

