% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propose.R
\name{fit_five}
\alias{fit_five}
\title{Helping to untangle the relationships between five populations based on given data.}
\usage{
fit_five(data, populations)
}
\arguments{
\item{data}{The data set.}

\item{populations}{A five element vector of population names.}
}
\value{
Printing stuff on console and "plots" tab. Making something fancy later.
}
\description{
Note that we do not say anything about the exact position of the root: the graphs
drawn are just chosen in a way they look nice. The first element of populations is
the leaf often drawn as the outgroup.
}

