% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propose.R
\name{fit_four}
\alias{fit_four}
\title{Helping to untangle the relationships between four populations based on given data.}
\usage{
fit_four(data, populations)
}
\arguments{
\item{data}{The data set.}

\item{populations}{A four element vector of population names.}
}
\value{
Printing stuff on console and "plots" tab. Making something fancy later.
}
\description{
Note that we do not say anything about the exact position of the root: the graphs
drawn are just chosen in a way they look nice. In all but one cases it is possible
to set the root on an outgroup branch, but sometimes the plot is so weird that we
nonetheless prefer to set it somewhere else. The first element of populations is
the leaf often drawn as the outgroup.
}

