% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_graph}
\alias{fit_graph}
\title{Fit the graph parameters to a data set.}
\usage{
fit_graph(data, graph, optimisation_options = NULL,
  parameters = extract_graph_parameters(graph))
}
\arguments{
\item{data}{The data set.}

\item{graph}{The admixture graph.}

\item{optimisation_options}{Options to the optimisation algorithm.}

\item{parameters}{In case one wants to tweak something in the graph.}
}
\value{
A list containing everything about the fit.
}
\description{
Tries to minimize the squared distance between statistics in \code{data} and
statistics given by the graph.
}
\details{
The data frame, \code{data}, must contain columns \code{W}, \code{X},
\code{Y}, and \code{Z}. The function then computes the \eqn{f_4(W,X;Y,Z)}
statistics for all rows from these to obtain the prediction made by the
graph.

The data frame must also contain a column, \code{D}, containing the
statistics observed in the data. The fitting algorithm attempts to minimize
the distance from this column and the predictions made by the graph.
}
\seealso{
\code{\link[neldermead]{optimset}}
}

