% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose.R
\name{fit_graph_list}
\alias{fit_graph_list}
\title{Fit lots of graphs to data.}
\usage{
fit_graph_list(data, graphs, cores)
}
\arguments{
\item{data}{The data table.}

\item{graphs}{List of graphs.}

\item{cores}{Number of cores used.}
}
\value{
A list of \code{\link{fast_fit}} results.
}
\description{
Fits a list of graphs to given data using parallel computation. This function
needs \code{doParallel}, \code{foreach} and \code{parallel} installed.
}
\seealso{
\code{\link{all_graphs}}

\code{\link{fit_permutations_and_graphs}}
}

