% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propose.R
\name{fit_permutations_and_graphs}
\alias{fit_permutations_and_graphs}
\title{Combine a list of permutations with a list of (parameterized) graphs to fit them}
\usage{
fit_permutations_and_graphs(data, permutations, graphs)
}
\arguments{
\item{permutations}{List of population permutations}

\item{graphs}{List of functions for producing graphs}
}
\value{
a list of fitted graphs.
}
\description{
Combine a list of permutations with a list of (parameterized) graphs to fit them
}

