% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitted.agraph_fit}
\alias{fitted.agraph_fit}
\title{Extract fitted data for a fitted graph.}
\usage{
\method{fitted}{agraph_fit}(object, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{...}{Additional parameters.}
}
\description{
Get the predicted f4 statistics for a fitted graph.
}

