% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose.R
\docType{data}
\name{five_leaves_graphs}
\alias{five_leaves_graphs}
\title{Five leaves graphs}
\format{An object of class \code{list} of length 8.}
\usage{
five_leaves_graphs
}
\value{
A list of functions on five leaves.
        The outputs of these functions are \code{\link{agraph}} objects.
}
\description{
A comprehensive listing of all the \eqn{8} admixture graphs with five leaves and
at most one admixture event. Our convention is that the position of the root does
not matter (as long as it's not after an admixture event) and that graphs that have
\emph{eyes}, two inner nodes with the property that all the paths between any two
leaves visits both or neither of them, are excluded. The reason is that the \eqn{f}
statistics can't detect the exact position of the root or distinguish between an eye
and a simple branch. The position of the root can be moved later with the function
\code{\link{make_an_outgroup}}.
}
\examples{
# While the usage of this function is pretty self-explanatory, let's plot all the graphs
# just for browsing.
for (i in seq(1, length(five_leaves_graphs))) {
  graph <- five_leaves_graphs[[i]](c("A", "B", "C", "D", "E"))
  # This is how you include quotation marks in strings by the way:
  title <- paste("five_leaves_graphs[[", i, "]](c(\\"A\\", \\"B\\", \\"C\\", \\"D\\", \\"E\\"))",
                 sep = "")
  plot(graph, color = "purple", title = title)
}

}
\seealso{
\code{\link{make_permutations}}

\code{\link{four_leaves_graphs}}

\code{\link{six_leaves_graphs}}

\code{\link{seven_leaves_trees}}

\code{\link{eight_leaves_trees}}

\code{\link{fit_permutations_and_graphs}}

\code{\link{add_a_leaf}}

\code{\link{add_an_admixture}}

\code{\link{add_an_admixture2}}

\code{\link{make_an_outgroup}}
}
\keyword{datasets}

