% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{format_path}
\alias{format_path}
\title{Create a path data frame from a list of nodes.}
\usage{
format_path(graph, nodes)
}
\arguments{
\item{graph}{The admixture graph the path is in.}

\item{nodes}{A list of nodes on a path.}
}
\value{
A data frame capturing the path and the probabilities/weights on the edges.
}
\description{
Creates a path data frame from a list of nodes.
}

