% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose.R
\name{graph_to_vector}
\alias{graph_to_vector}
\title{Graph to vector.}
\usage{
graph_to_vector(graph)
}
\arguments{
\item{graph}{A graph.}
}
\value{
The logical vector representing the graph.
}
\description{
Encodes an \code{\link{agraph}} object into a logical vector for saving memory. The admixture
proportion names will be lost.
}

