% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graphs_3_0}
\alias{graphs_3_0}
\title{Admixture graphs of 3 leaves and 0 admixture events compressed into vectors}
\format{A data frame with 1 row and 25 variables.}
\source{
Calculated using \code{\link{all_graphs}} and \code{\link{graph_to_vector}}.
}
\usage{
graphs_3_0
}
\description{
Use \code{\link{vector_to_graph}} on the \code{n}:th row to access the \code{n}:th graph
as an \code{\link{agraph}} object.
}
\keyword{datasets}

