% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graphs_5_1}
\alias{graphs_5_1}
\title{Admixture graphs of 5 leaves and 1 admixture event compressed into vectors}
\format{A data frame with 315 rows and 121 variables.}
\source{
Calculated using \code{\link{all_graphs}} and \code{\link{graph_to_vector}}.
}
\usage{
graphs_5_1
}
\description{
Use \code{\link{vector_to_graph}} on the \code{n}:th row to access the \code{n}:th graph
as an \code{\link{agraph}} object.
}
\keyword{datasets}

