% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graphs_8_0}
\alias{graphs_8_0}
\title{Admixture graphs of 8 leaves and 0 admixture events compressed into vectors}
\format{A data frame with 10395 rows and 225 variables.}
\source{
Calculated using \code{\link{all_graphs}} and \code{\link{graph_to_vector}}.
}
\usage{
graphs_8_0
}
\description{
Use \code{\link{vector_to_graph}} on the \code{n}:th row to access the \code{n}:th graph
as an \code{\link{agraph}} object.
}
\keyword{datasets}

