% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{is_descendant_of}
\alias{is_descendant_of}
\title{Is descendant of.}
\usage{
is_descendant_of(graph, offspring, ancestor)
}
\arguments{
\item{graph}{A graph.}

\item{offspring}{Potential offspring.}

\item{ancestor}{Potential ancestor.}
}
\value{
A truth value.
}
\description{
Tells whether a given node is descendant of another given node in a graph.
}

