% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signs.R
\name{is_negative}
\alias{is_negative}
\title{All overlaps are either empty or have a negative weight.}
\usage{
is_negative(overlaps)
}
\arguments{
\item{overlaps}{Data frame representing path overlaps, typically generated
by \code{\link{all_path_overlaps}}.}
}
\description{
All overlaps are either empty or have a negative weight.
}

