% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signs.R
\name{is_positive}
\alias{is_positive}
\title{All overlaps are either empty or have a positive weight.}
\usage{
is_positive(overlaps)
}
\arguments{
\item{overlaps}{Data frame representing path overlaps, typically generated
by \code{\link{all_path_overlaps}}.}
}
\description{
All overlaps are either empty or have a positive weight.
}

