% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signs.R
\name{is_unknown}
\alias{is_unknown}
\title{Overlapping edges have both positive and negative contributions.}
\usage{
is_unknown(overlaps)
}
\description{
Overlapping edges have both positive and negative contributions.
}

