% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signs.R
\name{is_zero}
\alias{is_zero}
\title{All overlaps are empty.}
\usage{
is_zero(overlaps)
}
\arguments{
\item{overlaps}{Data frame representing path overlaps, typically generated
by \code{\link{all_path_overlaps}}.}
}
\description{
All overlaps are empty.
}

