% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{make_mcmc_model}
\alias{make_mcmc_model}
\title{Collect the information about a graph and a data set needed to run an MCMC on it.}
\usage{
make_mcmc_model(graph, data)
}
\arguments{
\item{graph}{The admixture graph to analyse.}

\item{data}{The data set to compute the posterior over.}
}
\value{
A model object wrapping functions and data needed to sample from the MCMC.
}
\description{
Collect the information about a graph and a data set needed to run an MCMC on it.
}

