% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose.R
\name{make_permutations}
\alias{make_permutations}
\title{Make a list of all permutations of populations.}
\usage{
make_permutations(populations)
}
\description{
Right now it is hardwired to create permutations for four or five populations because
I just copied code from the two fit functions, but this should be generalized if we
need that at some point.
}

