% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{model_likelihood}
\alias{model_likelihood}
\title{Computes the likelihood of a model from samples from its posterior distribution.}
\usage{
model_likelihood(log_likelihoods)
}
\arguments{
\item{log_likelihoods}{Samples of log likelihoods from the posterior distribution of the graph.}
}
\value{
The likelihood of a graph where graph parameters are integrated out.
}
\description{
The likelihood of a graph can be computed by integrating over all the graph parameters (with appropriate priors).
Doing this by sampling from priors is very inefficient, so we use samples from the posteriors to importance
sample the likelihood.
}

