% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{mynonneg}
\alias{mynonneg}
\title{Non negative least square solution.}
\usage{
mynonneg(C, d)
}
\arguments{
\item{C}{The matrix.}

\item{d}{The vector.}
}
\value{
Look at the documentation of lsqnonneg.
}
\description{
This is lsqnonneg-function from the package pracma, just added some tolerance to
avoid crashing.
}

