% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-statistics.R
\name{no_admixture_events.agraph_fit}
\alias{no_admixture_events.agraph_fit}
\title{Get the number of admixture events in a fitted graph.}
\usage{
\method{no_admixture_events}{agraph_fit}(x)
}
\arguments{
\item{x}{The fitted graph.}
}
\value{
Number of admixture events in the graph.
}
\description{
Get the number of admixture events in a fitted graph.
}

