% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-statistics.R
\name{no_admixture_events.agraph_fit_list}
\alias{no_admixture_events.agraph_fit_list}
\title{Get the number of admixture events in a list of fitted graph.}
\usage{
\method{no_admixture_events}{agraph_fit_list}(x)
}
\arguments{
\item{x}{The graphs.}
}
\value{
Number of admixture events in the graphs.
}
\description{
Get the number of admixture events in a list of fitted graph.
}

