% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-statistics.R
\name{no_admixture_events}
\alias{no_admixture_events}
\title{Get the number of admixture events in a graph.}
\usage{
no_admixture_events(x)
}
\arguments{
\item{x}{The graph.}
}
\value{
Number of admixture events in the graph.
}
\description{
Get the number of admixture events in a graph.
}

