% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitted_plotting.R
\name{one_dimensional_plot}
\alias{one_dimensional_plot}
\title{A plot of the cost function around the best fit with respect to one admix variable specified
by the user.}
\usage{
one_dimensional_plot(object, X, resolution = 100, show_fit = FALSE,
  sigma = 6, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{X}{An admix variable name (remember quotation marks).}

\item{resolution}{How densely is the function evaluated.}

\item{show_fit}{Should the function plot the number of tests where the graph fits
the data instead of the sum of squared errors?}

\item{sigma}{If the function plots the fit, it is considered a hit if the difference
between a prediction and the observed statistics is no more than sigma/2.}

\item{...}{Additional parameters.}
}
\value{
Values for optimal cost function for values of X between zero and one, plotted.
}
\description{
Note that the other admix variables and the edge variables remain unchanged, so the plotted
value of the cost function is only optimal at the fit!
Sorry about the name, all the good ones were taken, and the fact that the word "graph" means
two different things does't help any.
}

