% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signs.R
\name{overlaps_sign}
\alias{overlaps_sign}
\title{Get the sign of overlapping paths.}
\usage{
overlaps_sign(overlaps)
}
\arguments{
\item{overlaps}{Data frame representing path overlaps, typically generated
by \code{\link{all_path_overlaps}}.}
}
\description{
Get the sign of overlapping paths.
}

