% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agraph.R
\name{parent_edges}
\alias{parent_edges}
\title{Create the list of edges for an admixture graph.}
\usage{
parent_edges(edges)
}
\arguments{
\item{edges}{List of edges.}
}
\description{
Syntactic suggar for constructing edges in an admixture graph.
}

