% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.agraph}
\alias{plot.agraph}
\title{Plot an admixture graph.}
\usage{
\method{plot}{agraph}(x, ordered_leaves = NULL,
  show_admixture_labels = FALSE, show_inner_node_labels = FALSE, ...)
}
\arguments{
\item{x}{The admixture graph.}

\item{ordered_leaves}{The leaf-nodes in the left to right order they should
be drawn. I don't have a good algorithm for figuring out that order so for
now it is required as a function argument.}

\item{show_admixture_labels}{A flag determining if the plot should include
the names of admixture proportions.}

\item{show_inner_node_labels}{A flat determining if the plot should include
the names of inner nodes.}

\item{...}{Additional plotting options}
}
\description{
This is a basic drawing routine for visualising the graph. For publication
quality graphs a lot more tweaking is probably needed.
}

