% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitted_plotting.R
\name{plot.agraph_fit}
\alias{plot.agraph_fit}
\title{Plot the fit of a graph to data.}
\usage{
\method{plot}{agraph_fit}(x, sigma = 6, ...)
}
\arguments{
\item{x}{Fitted graph object.}

\item{sigma}{How many sigmas the error bars should be wide.}

\item{...}{Additional parameters.}
}
\description{
Plot the fit of a graph to data.
}

