% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propose.R
\name{plot.agraph_fit_list}
\alias{plot.agraph_fit_list}
\title{Plot a list of fitted admixture graphs.}
\usage{
\method{plot}{agraph_fit_list}(x, measure = "SSE", ...)
}
\arguments{
\item{x}{List of fitted graphs.}

\item{...}{Additional plotting options}
}
\description{
Plot a list of fitted admixture graphs.
}

