% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4_plotting.R
\name{plot.f4stats}
\alias{plot.f4stats}
\title{Plot the fit of a graph to data.}
\usage{
\method{plot}{f4stats}(x, sigma = 6, ...)
}
\arguments{
\item{x}{Data frame with observed \eqn{D} (\eqn{f_4}) statistics}

\item{sigma}{How many sigmas the error bars should be wide.}

\item{...}{Additional parameters.}
}
\description{
Plot the fit of a graph to data.
}

