% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary-statistics.R
\name{poor_fits.agraph_fit}
\alias{poor_fits.agraph_fit}
\title{Get the tests in the fit where the predictions fall outside of the error bars.}
\usage{
\method{poor_fits}{agraph_fit}(fit, sigma = 6)
}
\arguments{
\item{fit}{The fitted graph}

\item{sigma}{the width of the error bars.}
}
\value{
The poorly fitted tests.
}
\description{
Get the tests in the fit where the predictions fall outside of the error bars.
}

