% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-statistics.R
\name{poor_fits}
\alias{poor_fits}
\title{Get the tests in the fit where the predictions fall outside of the error bars.}
\usage{
poor_fits(fit, sigma = 6)
}
\arguments{
\item{fit}{The fitted graph.}

\item{sigma}{The width of the error bars.}
}
\value{
The poorly fitted tests.
}
\description{
Get the tests in the fit where the predictions fall outside of the error bars.
}

