% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{predict.agraph_fit}
\alias{predict.agraph_fit}
\title{Predict statistics on new data.}
\usage{
\method{predict}{agraph_fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{newdata}{New data frame to predict values for.}

\item{...}{Additional arguments.}
}
\description{
Predict expected f4 statistics. If \code{newdata} is not specified this function just
returns the predicted values on the original data.
}

