% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{print.agraph_fit}
\alias{print.agraph_fit}
\title{Print function for a fitted graph.}
\usage{
\method{print}{agraph_fit}(object)
}
\arguments{
\item{object}{The fitted object.}
}
\description{
Print summary of the result of a fit.
}

