% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{residuals.agraph_fit}
\alias{residuals.agraph_fit}
\title{Extract the individual errors in a fitted graph.}
\usage{
\method{residuals}{agraph_fit}(object, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{...}{Additional parameters.}
}
\description{
Get D - graph_f4 for each data point used in the fit.
}

