% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_metropolis_hasting}
\alias{run_metropolis_hasting}
\title{Run a Metropolis-Hasting MCMC to sample graph parameters.}
\usage{
run_metropolis_hasting(model, initial_state, iterations)
}
\arguments{
\item{model}{Object constructed with \code{\link{make_mcmc_model}}.}

\item{initial_state}{The initial set of graph parameters.}

\item{iterations}{Number of iterations to sample.}
}
\value{
A matrix containing the trace of the MCMC run.
}
\description{
Run a Metropolis-Hasting MCMC to sample graph parameters.
}

