% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic-f-statistics.R
\name{sf3}
\alias{sf3}
\title{Calculate the f_3(A; B, C) statistics.}
\usage{
sf3(graph, A, B, C)
}
\arguments{
\item{graph}{The admixture graph.}

\item{A}{A leaf node.}

\item{B}{A leaf node.}

\item{C}{A leaf node.}
}
\value{
A symbolic representation of the equation for the \eqn{f_3}
        statistics given by the admixture graph.
}
\description{
Calculate the \eqn{f_3(A; B, C)} statistics.
}

