% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic-f-statistics.R
\name{sf4}
\alias{sf4}
\title{Calculate the f_4(W, X; Y, Z) statistics.}
\usage{
sf4(graph, W, X, Y, Z)
}
\arguments{
\item{graph}{The admixture graph.}

\item{W}{A leaf node.}

\item{X}{A leaf node.}

\item{Y}{A leaf node.}

\item{Z}{A leaf node.}
}
\value{
A symbolic representation of the equation for the \eqn{f_4}
        statistics given by the admixture graph.
}
\description{
Calculate the \eqn{f_4(W, X; Y, Z)} statistics.
}

