% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_and_split.R
\name{split_population.agraph_fit}
\alias{split_population.agraph_fit}
\title{Reverse a projection of samples to populations.}
\usage{
\method{split_population}{agraph_fit}(x)
}
\arguments{
\item{x}{The projected data or a fitted object on projected data.}
}
\description{
Reverse a projection of samples to populations.
}

