% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_and_split.R
\name{split_population}
\alias{split_population}
\title{Reverse a projection of samples to populations.}
\usage{
split_population(x)
}
\arguments{
\item{x}{The projected data or a fitted object on projected data.}
}
\description{
Reverse a projection of samples to populations.
}

