% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-statistics.R
\name{sum_of_squared_errors.agraph_fit}
\alias{sum_of_squared_errors.agraph_fit}
\title{Get the sum of squared errors for a fitted graph.}
\usage{
\method{sum_of_squared_errors}{agraph_fit}(x)
}
\arguments{
\item{x}{The fitted graph.}
}
\value{
SSE for the fit.
}
\description{
Get the sum of squared errors for a fitted graph.
}

