% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-statistics.R
\name{sum_of_squared_errors.agraph_fit_list}
\alias{sum_of_squared_errors.agraph_fit_list}
\title{Get the sum of squared errors for a list of fitted graph.}
\usage{
\method{sum_of_squared_errors}{agraph_fit_list}(x)
}
\arguments{
\item{x}{The fitted graphs.}
}
\value{
SSE for the fits.
}
\description{
Get the sum of squared errors for a list of fitted graph.
}

