% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{summary.agraph_fit}
\alias{summary.agraph_fit}
\title{Print function for a fitted graph.}
\usage{
\method{summary}{agraph_fit}(object, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{...}{Additional parameters.}
}
\description{
Print summary of the result of a fit.
}

