% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propose.R
\name{summary.agraph_fit_list}
\alias{summary.agraph_fit_list}
\title{Print function for the fitted graphs.}
\usage{
\method{summary}{agraph_fit_list}(object, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{...}{Additional parameters.}
}
\description{
Print summary of the result of a list of fitted graphs.
}

