% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{thinning}
\alias{thinning}
\title{Thins out an MCMC trace.}
\usage{
thinning(trace, k)
}
\arguments{
\item{trace}{A trace from an MCMC run.}

\item{k}{The number of lines to skip over per retained sample.}
}
\description{
Thins out an MCMC trace.
}

