% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose.R
\name{vector_to_graph}
\alias{vector_to_graph}
\title{Vector to graph.}
\usage{
vector_to_graph(vector)
}
\arguments{
\item{vector}{A logical vector.}
}
\value{
The graph corresponding to the vector.
}
\description{
Interprets a logical vector back to an \code{\link{agraph}} object. The admixture
proportion names are now lost.
}

