% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agraph.R
\name{admixture_edge}
\alias{admixture_edge}
\title{Create an admixture edge from a child to two parents.}
\usage{
admixture_edge(child, parent1, parent2, prop = NA)
}
\arguments{
\item{child}{The name of the child node.}

\item{parent1}{The name of the parent node.}

\item{parent2}{The name of the parent node.}

\item{prop}{Admixture proportions from \code{parent1} to \code{child}.
If this parameter is not provided, you must explicitly
specify the admixture proportion parameters in the
\code{agraph} function call.}
}
\description{
Syntactic suggar for constructing edges in an admixture graph.
}

