% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agraph.R
\name{agraph_parents}
\alias{agraph_parents}
\title{Build the parent incidence matrix from an edge list.}
\usage{
agraph_parents(nodes, parent_edges)
}
\arguments{
\item{nodes}{Nodes in the admxture graph.}

\item{parent_edges}{An \eqn{n \times 2}{n x 2} matrix where the first column is the
child node and the second is the parent.}
}
\value{
An incidence matrix for the parent structure of an admixture graph.
}
\description{
Build the parent incidence matrix from an edge list.
}
\seealso{
\code{\link{agraph_children}}
}

