% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agraph.R
\name{agraph_weights}
\alias{agraph_weights}
\title{Build the matrix of admixture proportions from an edge list.}
\usage{
agraph_weights(nodes, admixture_weights, parents)
}
\arguments{
\item{nodes}{The name of the nodes in the admxture graph.}

\item{admixture_weights}{An \eqn{n \times 3}{n x 3} matrix where the first column is
the child node, the second isthe parent and the third
is the admixture weight on that edge.}

\item{parents}{The parent edge list. Used for checking graph consistency.}
}
\value{
A matrix containing the admixture weights.
}
\description{
Build the matrix of admixture proportions from an edge list.
}

