% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{build_edge_optimisation_matrix}
\alias{build_edge_optimisation_matrix}
\title{Build a matrix coding the linear system of edges once the admix variables
have been fixed.}
\usage{
build_edge_optimisation_matrix(data, graph,
  parameters = extract_graph_parameters(graph))
}
\arguments{
\item{data}{The data set.}

\item{graph}{The admixture graph.}

\item{parameters}{In case one wants to tweak something in the graph.}
}
\value{
A list containing the full matrix (\code{$full}), a version with zero
        columns removed (\code{$column_reduced}), a version with zero rows and
        repeated rows also removed (\code{$double_reduced}), and an indicator
        of warning (\code{$complaint}).
}
\description{
The elements are characters containing numerals, admix variable names,
parenthesis and arithmetical operations. (Transform into expressions with
\code{parse} and then evaluate with \code{eval}). The column names are the
edge names from \code{extract_graph_parameter$edges}, the rows have no names.
}
\details{
If the essential number of equations is not higher than the essential number of
edge variables, the quality of edge optimisation will not depend on the admix
variables (expect in a very special cases), and a complaint will be given.
}

