% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{coef.agraph_fit}
\alias{coef.agraph_fit}
\title{Parameters for the fitted graph.}
\usage{
\method{coef}{agraph_fit}(object, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{...}{Additional parameters.}
}
\description{
Extracts the graph parameters for the graph fitted to data. Note that the optimal 
parameters are generally not unique.
}
\seealso{
\code{link{summary.agraph_fit}}
}

