% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_plotting.R
\name{contour_plot}
\alias{contour_plot}
\title{A contour plot of the cost function around the best (admix variable) fit with respect to 2
admix variables specified by the user.}
\usage{
contour_plot(object, X, Y, resolution = 10, show_fit = FALSE, sigma = 6,
  ...)
}
\arguments{
\item{object}{The fitted object.}

\item{X}{An admix variable name (remember quotation marks).}

\item{Y}{An admix variable name (remember quotation marks).}

\item{resolution}{How densely is the function evaluated.}

\item{show_fit}{Should the function plot the number of tests where the graph fits
the data instead of the sum of squared errors?}

\item{sigma}{If the function plots the fit, it is considered a hit if the difference
between a prediction and the observed statistics is no more than sigma/2.}

\item{...}{Additional parameters passed to the plotting function filled.contour.}
}
\value{
The matrix of values computed and plotted.
}
\description{
Note that the other admix variables remain unchanged, so the plotted value of the cost
function might only be optimal in one point if there are more than 2 admix variables!
The edge variables are re-optimized at every pont.
}

