% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitted_plotting.R
\name{contour_plot}
\alias{contour_plot}
\title{A contour plot of the cost function around the best fit with respect to 2
admix variables specified by the user.}
\usage{
contour_plot(object, X, Y, resolution = 10, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{X}{An admix variable name (remember quotation marks) or number.}

\item{Y}{An admix variable name (remember quotation marks) or number.}

\item{resolution}{How densely is the function evaluated.}

\item{...}{Additional parameters passed to the plotting function filled.contour}
}
\value{
Just a contour plot with FIERY colours.
}
\description{
Returning the cost, an example edge solution of an optimal fit, and linear
relations describing the set of all edge solutions. Operating with the full
edge optimisation matrix.
}

