% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{cost_function}
\alias{cost_function}
\title{The cost function fed to nelder mead.}
\usage{
cost_function(data, matrix, graph,
  parameters = extract_graph_parameters(graph))
}
\arguments{
\item{data}{The data set.}

\item{matrix}{A column reduced edge optimisation matrix (typically given by
the function \code{edge_optimisation_matrix$column_reduced}).}

\item{graph}{The admixture graph.}

\item{parameters}{In case one wants to tweak something in the graph.}
}
\value{
Given an input vector of admix variables, returns the smallest error
         regarding the edge variables.
}
\description{
We want nelder mead to run fast so the cost function operates with the column
rduced edge optimisation matrix and does not give any extar information about
the fit. For the details, use \code{edge_optimisation_function} instead.
}

