% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitted.agraph_fit}
\alias{fitted.agraph_fit}
\title{Predicted f statistics for the fitted graph.}
\usage{
\method{fitted}{agraph_fit}(object, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{...}{Additional parameters.}
}
\description{
Gets the predicted \eqn{f} statistics \eqn{F} for the fitted graph.
}
\seealso{
\code{link{summary.agraph_fit}}
}

