% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signs.R
\name{is_unknown}
\alias{is_unknown}
\title{Overlapping edges have both positive and negative contributions.}
\usage{
is_unknown(overlaps)
}
\arguments{
\item{overlaps}{Data frame representing path overlaps, typically generated
by \code{\link{all_path_overlaps}}.}
}
\description{
Overlapping edges have both positive and negative contributions.
}

