% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{mynonneg}
\alias{mynonneg}
\title{Non negative least square solution.}
\usage{
mynonneg(C, d, iteration_multiplier = 3)
}
\arguments{
\item{C}{The matrix.}

\item{d}{The vector.}

\item{iteration_multiplier}{The definition of "too many steps".}
}
\value{
A vector ($x) and the error ($resid.norm).
}
\description{
This is lsqnonneg-function from the package pracma, just changed qr.solve into
using Moore-Penrose inverse instead (ginv from MASS) as qr.solve crashes for
some singular inputs. Now it won't crash but it's sometimes running for very long
time (forever?), presumably with those problematic inputs. After too many steps
the function halts and lies that the fit was terrible. I don't think this will
cause problems.
}

