% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{print.agraph_fit}
\alias{print.agraph_fit}
\title{Print function for the fitted graph.}
\usage{
\method{print}{agraph_fit}(x, ...)
}
\arguments{
\item{x}{The fitted object.}

\item{...}{Additional parameters.}
}
\description{
Prints the value of \code{\link{cost_function}} of the fitted graph, and complains
if some or all of the admixture proportions aren't trurly fitted.
Note: the admixture proportion not being trurly fitted does not necessarily mean that
there is no evidence of an admix event!
}
\seealso{
\code{link{summary.agraph_fit}}
}

