% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{residuals.agraph_fit}
\alias{residuals.agraph_fit}
\title{Errors of prediction in the fitted graph}
\usage{
\method{residuals}{agraph_fit}(object, ...)
}
\arguments{
\item{object}{The fitted object.}

\item{...}{Additional parameters.}
}
\description{
Gets \eqn{f - F}, the difference between predicted and observed  statistics,
for each data point used in the fit.
}
\seealso{
\code{link{summary.agraph_fit}}
}

