% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose.R
\docType{data}
\name{seven_leaves_trees}
\alias{seven_leaves_trees}
\title{Seven leaves trees.}
\format{A list of functions on seven leaves.
        The outputs of these functions are \code{\link{agraph}} objects.}
\usage{
seven_leaves_trees
}
\description{
The function \code{\link{seven_leaves_graphs}} is better than this as it also contains
graphs with one admixture. (This function is only kept for legacy reasons.)
Even more obsolete since the introduction of \code{\link{all_graphs}}.
}
\keyword{datasets}

